+++ app/Livewire/Profile/Settings.php
@@ -9,8 +9,19 @@ use Livewire\Component;
 
 class Settings extends Component
 {
+    public string $customLinkName;
+    public string $customLink;
+
+    public function mount(): void
+    {
+        $this->customLinkName =
+            Configuration::getString(ConfigurationKey::NAVIGATION_FAV_CUSTOM_LINK_NAME, auth()->user(), '');
+        $this->customLink =
+            Configuration::getString(ConfigurationKey::NAVIGATION_FAV_CUSTOM_LINK, auth()->user(), '');
+    }
+
     #[Renderless]
-    public function dashboardButtonChangedBirthdayList(bool $enabled): void
+    public function customNavBirthdayListBtnChanged(bool $enabled): void
     {
         Configuration::storeBool(ConfigurationKey::NAVIGATION_FAV_BIRTHDAY_LIST, $enabled, auth()->user());
     }
@@ -19,6 +30,16 @@ class Settings extends Component
     public function customNavLinkEnabledChanged(bool $enabled): void
     {
         Configuration::storeBool(ConfigurationKey::NAVIGATION_FAV_CUSTOM_LINK_ENABLED, $enabled, auth()->user());
+    }
+
+    public function updatedCustomLinkName(): void
+    {
+        Configuration::storeString(ConfigurationKey::NAVIGATION_FAV_CUSTOM_LINK_NAME, $this->customLinkName, auth()->user());
+    }
+
+    public function updatedCustomLink(): void
+    {
+        Configuration::storeString(ConfigurationKey::NAVIGATION_FAV_CUSTOM_LINK, $this->customLink, auth()->user());
     }
 
     public function render(): \Illuminate\Contracts\View\View|\Illuminate\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\Foundation\Application